a little bit edited code

+++ src/main/java/by/yurachel/web_app/servlet/ChangeServlet.java
@@ -25,7 +25,7 @@ public class ChangeServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         String oldPhoneName = req.getParameter("oldName");
-        int oldPhoneId = pr.findPhoneIDByName(oldPhoneName);
+        long oldPhoneId = pr.findPhoneIDByName(oldPhoneName);
         Phone newPhone = new Phone(oldPhoneId, req.getParameter("name"), Double.parseDouble(req.getParameter("price")), req.getParameter("processor"));
         ROOT_LOGGER.info("Old phone {} was changed by new phone {}", oldPhoneName, newPhone);
         pr.changePhoneParam(newPhone, oldPhoneName);
